edit .proto

+++ ServerLib/User.h
@@ -3,15 +3,15 @@
 
 #pragma pack(push, 1)
 struct Position {
-	int32 x = 0;
-	int32 y = 0;
-	int32 z = 0;
+	float x = 0;
+	float y = 0;
+	float z = 0;
 };
 
 struct Direction {
-	int32 x = 0;
-	int32 y = 0;
-	int32 z = 0;
+	float x = 0;
+	float y = 0;
+	float z = 0;
 };
 #pragma pack(pop)
 
@@ -28,8 +28,8 @@ public:
 	int64 GetLastMovePacket() { return _lastMovePacket; }
 
 	void SetId(int32 id) { _id = id; }
-	void SetPosition(int32 x, int32 y, int32 z) { _position.x = x; _position.y = y; _position.z = z; }
-	void SetDirection(int32 x, int32 y, int32 z) { _direction.x = x; _direction.y = y; _direction.z = z; }
+	void SetPosition(float x, float y, float z) { _position.x = x; _position.y = y; _position.z = z; }
+	void SetDirection(float x, float y, float z) { _direction.x = x; _direction.y = y; _direction.z = z; }
 	void SetLastMovePacket(int64 time) { _lastMovePacket = time; }
 
 private:
@@ -38,7 +38,7 @@ private:
 	int32 _id = 0;
 	Position _position = {};
 	Direction _direction = {};
-	int32 _moveSpeed = 0;
+	float _moveSpeed = 0;
 	int64 _lastMovePacket = 0;
 
 };
